[FLINK-16396] [kafka] Rename KafkaSpecJsonParser to KafkaIngressSpecJsonParser
To prepare adding a YAML-ized Kafka egress, this commit renames the
KafkaSpecJsonParser to mention "Ingress" as there would be a counterpart
to that for the egress.
